==> 2_Set_Project.ipynb <==
## Generate project configure file

import numpy as np
import os
import shutil
import yaml

name_project = 'output_regular_v9.5_all/'
rdir_CC = 'CC/CC_150_prewhiten/'  

# 找到name_project中的第二个'_'的index
tag = name_project[ name_project.rfind('_')+1: -1]
rdir_project = 'project/' + name_project
tag

proj = {}
proj['name'] = rdir_project
with open('a-project.yml', 'w') as f:
    yaml.dump(proj, f)

with open('0_config.yml', 'r', encoding='utf-8') as f:
    dir_config = yaml.load(f.read(), Loader=yaml.FullLoader)
dir_project_workspace = dir_config['dir_project_workspace']
dir_CC_workspace = dir_config['dir_CC_workspace']
print('dir_CC_workspace: ', dir_CC_workspace)
print('dir_project_workspace: ', dir_project_workspace)
dir_project = dir_project_workspace + rdir_project
print('dir_project: ', dir_project)

dir_CC = dir_CC_workspace + rdir_CC
filename = dir_CC+'info_CC.npy'
info_CC = np.load(filename, allow_pickle='TRUE').item()

info_CC

### Basic settings

# set SAC data dir ,correspoding stalistname, and output dir
info = {}
info_bi = {}
#filename = dir_project+'Basic_info.npy'
#info = np.load(filename, allow_pickle='TRUE').item()

# SAC dir
info['rdir_project'] = name_project
info['tag'] = tag
info['stalistname_all'] = info_CC['stalistname']
# correlation all dirtin
info['rdir_CC'] = rdir_CC
# config dir
info['rdir_image']  = 'image_'+tag + '/'

info['nf'] = info_CC['nf']
info['Fs'] = info_CC['Fs']
info_bi['f'] = info_CC['f']
f = info_CC['f']
dt = 1/np.max(f)
t = (np.linspace(-len(f)-1,len(f)-1,2*(len(f)-1))+0.5)*dt/2
info_bi['t'] = t

### Generate folders 

if os.path.exists(dir_project):
    shutil.rmtree(dir_project)
    os.mkdir(dir_project)
else:
    os.mkdir(dir_project)
if not os.path.exists(dir_project+info['rdir_image']):
    os.mkdir(dir_project + info['rdir_image'])

np.save(dir_project+'/Basic_info.npy',info_bi)
with open(dir_project+'Basic_info.yml', 'w', encoding='utf-8') as f:
   yaml.dump(data=info, stream=f, allow_unicode=True)
