==> p_6_compare-FJs.ipynb <==
import numpy as np
import os
import matplotlib.pyplot as plt
import h5py
from scipy import interpolate
from ccfj import GetStationPairs
import pandas as pd
import yaml
import geopandas as gp
import time

import sys
sys.path.append(r'../tools_F-J/')
from toollib_standard import maplib
from toollib_standard import mathlib
from toollib_standard import filelib
from toollib_standard import stacklib
from toollib_standard import plotlib

with open('0_config.yml', 'r', encoding='utf-8') as f:
    dir_config = yaml.load(f.read(), Loader=yaml.FullLoader)
dir_project_workspace = dir_config['dir_project_workspace']
dir_CC_workspace = dir_config['dir_CC_workspace']
print('dir_CC_workspace: ', dir_CC_workspace)
print('dir_project_workspace: ', dir_project_workspace)

"""
name_projects = []
name_projects.append('project/output_FJSJ_17-02/')
name_projects.append('project/output_FJSJ_17-03/')
name_projects.append('project/output_FJSJ_17-04/')
name_projects.append('project_repartition_v3.0/output_repar_v9.1_02-01/')
name_projects.append('project_repartition_v4.0/output_repar_v9.2_01-01/')
"""
name_projects = []
name_projects.append('project/output_regular_v9.5_3x7/')
name_projects.append('project/output_regular_v9.5_5x13/')
name_projects.append('project/output_regular_v9.5_7x19/')
name_projects.append('project_repartition_v4.0/output_repar_v9.5_01--10-16Hz/')

name_project_sample = 'project_repartition_v4.0/output_repar_v9.5_01--10-16Hz/'
dir_project_sample = dir_project_workspace + name_project_sample
filename = dir_project_sample+'Basic_info.yml'
with open(filename, 'r', encoding='utf-8') as f:
    info_basic = yaml.load(f.read(), Loader=yaml.FullLoader)
key_subworks = info_basic['key_subworks']
key_subworks

stalistname_all = info_basic['stalistname_all']
stainfo = pd.read_excel(stalistname_all)
nsta_all = len(stainfo.iloc[:,0])
StationPairs_all = GetStationPairs(nsta_all)
nPairs_all = int(len(StationPairs_all)/2)
stalist_all = stainfo['Station'].tolist()
lat_stations_all = stainfo['latitude'].tolist() 
lon_stations_all = stainfo['longitude'].tolist()

stations_alls = {}
lat_stations_alls = {}
lon_stations_alls = {}
lat_centroid_alls = {}
lon_centroid_alls = {}
for name_project in name_projects:
    dir_project = dir_project_workspace + name_project
    filename = dir_project+'Basic_info.yml'
    with open(filename, 'r', encoding='utf-8') as f:
        info_basic = yaml.load(f.read(), Loader=yaml.FullLoader)
    dir_partition = dir_project + 'partition/'
    stations = {}
    lat = {}
    lon = {}
    lat_centroid = {}
    lon_centroid = {}
    for key in info_basic['key_subworks']:
        filepath = dir_partition + str(key) + '.txt'
        stations_this, lat_stations_this, lon_stations_this = np.loadtxt(filepath, dtype='str' , unpack=True)
        stations[key] = stations_this.tolist()
        lat[key] = lat_stations_this.astype(float)
        lon[key] = lon_stations_this.astype(float)
        lat_centroid[key] = np.mean(lat[key])
        lon_centroid[key] = np.mean(lon[key])

    stations_alls[name_project] = stations
    lat_stations_alls[name_project] = lat
    lon_stations_alls[name_project] = lon
    lat_centroid_alls[name_project] = lat_centroid
    lon_centroid_alls[name_project] = lon_centroid
    

faults = np.load('clark_faults.npy', allow_pickle='TRUE').item()

dir_image = dir_project_workspace + 'compare_relative_error/'
if not os.path.exists(dir_image):
    os.makedirs(dir_image)

def find_nearset(key_subwork,name_project):
	global name_project_sample
	global key_subworks

	lon_this = lon_centroid_alls[name_project_sample][key_subwork]
	lat_this = lat_centroid_alls[name_project_sample][key_subwork]

	lon_stations_all = []
	lat_stations_all = []
	
	lon_stations_all = list(lon_centroid_alls[name_project].values())
	lat_stations_all = list(lat_centroid_alls[name_project].values())
	key_all = list(lat_centroid_alls[name_project].keys())

	dist = np.sqrt((np.array(lon_stations_all)-lon_this)**2+(np.array(lat_stations_all)-lat_this)**2)
	indx = np.argsort(dist)
	#print(indx)
	key_nearset = key_all[indx[0]]

	return str(key_nearset)

c_map = 'jet'
v_min = 0.1
v_max = None
xlim_f = [1,25]
for key_subwork in key_subworks:
    print('start ',str(key_subwork),' ',str(key_subworks.index(key_subwork)+1),'/',len(key_subworks))
    fig,ax = plt.subplots(2,len(name_projects),figsize=(5*len(name_projects),9))
    for name_project in name_projects:
        key_this_find = find_nearset(key_subwork,name_project)
        #print(name_project)

        stations = stations_alls[name_project]
        lat_stations = lat_stations_alls[name_project]
        lon_stations = lon_stations_alls[name_project]


        col_index = name_projects.index(name_project)
        dir_project = dir_project_workspace + name_project
        filename = dir_project+'Basic_info.yml'
        with open(filename, 'r', encoding='utf-8') as f:
            info_basic = yaml.load(f.read(), Loader=yaml.FullLoader)
        dir_ds = dir_project + info_basic['rdir_ds']

        #print('key_this_find: ',key_this_find)

        stalist = stations[key_this_find]
        lat_this = lat_stations[key_this_find]
        lon_this = lon_stations[key_this_find]
        nsta = len(stalist)

        ax[1][col_index]=plotlib.plot_area(ax[1][col_index],lon_stations_all,lat_stations_all,lon_this,lat_this)
        for i in range(len(faults)):
            ax[1][col_index].plot(faults['clark'+str(i+1)]['lon'], faults['clark'+str(i+1)]['lat'], 'k')
        ax[1][col_index].scatter(np.mean(lon_this),np.mean(lat_this),c='g',marker = '^',s=50)
        
        title0 = name_project.split('/')[-2] + ' ' +str(key_this_find)
        c = np.linspace(info_basic['fj_c_min'],info_basic['fj_c_max'],info_basic['fj_c_num'])
        ds = h5py.File(dir_ds+'ds_'+str(key_this_find)+'.h5', 'r')
        ds_remove = ds['ds_remove'][0]
        ds_linear = ds['ds_linear'][0]
        ds_linear = plotlib.smooth_ds(ds_linear)
        ds_remove = plotlib.smooth_ds(ds_remove)
        f = ds['f'][:]
        ax[0][col_index] = plotlib.plot_fj(ax[0][col_index] ,ds_remove,title0,f,c,0,v_min=v_min,v_max=v_max,c_map=c_map)
        ax[0][col_index] .set_xlim(xlim_f)
        #ax[1][col_index] = plotlib.plot_fj(ax[1][col_index] ,ds_linear,title0,f,c,0,v_min=v_min,v_max=v_max,c_map=c_map)
        #ax[1][col_index] .set_xlim(xlim_f)
    plt.tight_layout()
    plt.savefig(dir_image+key_subwork+'.png',dpi=60)
    plt.close()
    #time.sleep(7)
